+++ Steps/Communications/Http/src/SendHttpRequestOptions.cs
@@ -1,7 +1,7 @@
 ﻿// <copyright file="SendHttpRequestOptions.cs" company="Chris Trout">
 // MIT License
 //
-// Copyright(c) 2021 Chris Trout
+// Copyright(c) 2021-2022 Chris Trout
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -28,7 +28,6 @@ namespace MyTrout.Pipelines.Steps.Communications.Http
     using MyTrout.Pipelines.Core;
     using System;
     using System.Collections.Generic;
-    using System.ComponentModel;
     using System.Net.Http;
 
     /*
@@ -41,6 +40,11 @@ namespace MyTrout.Pipelines.Steps.Communications.Http
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
     public class SendHttpRequestOptions
     {
+        /// <summary>
+        /// Gets or sets the HTTP Content-Type Content Header that is specified when <see cref="UploadInputStreamAsContent"/> is <see langword="true"/>.
+        /// </summary>
+        public string ContentTypeHeaderValue { get; set; } = "application/json";
+
         /// <summary>
         /// Gets or sets the HTTP Headers from <see cref="IPipelineContext.Items"/> that need to be set for the <see cref="HttpEndpoint" /> to work.
         /// </summary>
@@ -90,9 +94,14 @@ namespace MyTrout.Pipelines.Steps.Communications.Http
         public string IsSuccessfulStatusCodeContextName { get; set; } = HttpCommunicationConstants.HTTP_IS_SUCCESSFUL_STATUS_CODE;
 
         /// <summary>
-        /// Gets the HTTP Method value from <see cref="IConfiguration"/>.
+        /// Gets or sets the HTTP Method value from <see cref="IConfiguration"/>.
         /// </summary>
-        public string HttpMethod { get; set; }
+        public string HttpMethod { get; set; } = "POST";
+
+        /// <summary>
+        /// Gets or sets a list of values that will be replaced in the <see cref="HttpEndpoint"/> from the <see cref="IPipelineContext"/>.
+        /// </summary>
+        public List<string> HttpEndpointReplacementKeys { get; set; } = new List<string>();
 
         /// <summary>
         /// Gets or sets the Output Stream Context Name used for <see cref="IPipelineContext.Items"/>.